requirements.txt
flask>=2.0.0
flask-socketio>=5.0.0
opencv-python-headless>=4.5.0,<4.9.0
pillow>=8.0.0,<11.0.0
numpy<2.0,>=1.21.0
scipy<1.14,>=1.7.0
mediapipe>=0.8.9,<0.11.0
torch==1.13.1
torchvision==0.14.1
basicsr==1.4.2
facexlib==0.3.0
gfpgan==1.3.8
onnxruntime>=1.10.0,<1.17.0
realesrgan>=0.3.0
wget>=3.2
werkzeug>=2.0.0,<3.0.0
stripe>=5.0.0
flask-mail>=0.9.1
python-dotenv>=1.0.0
sib-api-v3-sdk>=7.6.0